requirements.txt
click==7.1.2
GitPython==3.1.14
pyotp==2.6.0
PyYAML==5.4.1
